Add files via upload
Dubai Real Estate Market Analysis and Predictive Modeling

This project explores and analyzes the Dubai real estate market, using a combination of machine learning models and data visualization techniques. The goal is to predict rental prices and identify the most profitable property types across different locations in Dubai.

Key Components:

    Data Exploration: In-depth analysis of the dataset to understand the relationships between various factors influencing rental prices.
    Predictive Modeling: Development of multiple models, including Multiple Linear Regression, Random Forest, and Neural Networks, to forecast rental prices.
    Geospatial Analysis: Creation of an interactive map using folium to visualize the most profitable property types in each region.
    Tools and Libraries: The project utilizes pandas, scikit-learn, TensorFlow, and folium to perform data processing, model building, and visualization.

This project provides a comprehensive overview of the potential applications of data science in the real estate sector, emphasizing the importance of understanding data deeply and applying creativity in analysis.

Feedback and contributions are welcome!